Refactor license management into separate module with improved functions
